Remove up arrow icon FOUC (admin bar)
When the site first loads, the css that sets this value isn't available. As a result the initial 512x512 svg loads causing a large arrow icon to briefly appear on each payload (logged in users, of course).

Let's set the w/h properties at the svg level to prevent this.